requirements.txt
APScheduler           #3.9.1
bottle                #0.12.21
certifi               #2022.5.18.1
cffi                  #1.15.0
charset-normalizer    #2.0.12
cryptography          #37.0.2
dict2xml              #1.7.1
idna                  #3.3
numpy                 #1.22.4
pandas                #1.4.2
Paste                 #3.5.0
pip                   #21.2.4
pycparser             #2.21
PyJWT                 #2.4.0
python-dateutil       #2.8.2
pytz                  #2022.1
pytz-deprecation-shim #0.1.0.post0
PyYAML                #6.0
requests              #2.27.1
schedule              #1.1.0
setuptools            #58.1.0
six                   #1.16.0
tzdata                #2022.1
tzlocal               #4.2
urllib3               #1.26.9